pyproject.toml
[tool.poetry]
name = "django-struct"
version = "0.1.0"
description = ""
authors = ["Yaroslav <yaspichihin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "=3.11.9"
django = "=5.0.4"
django-environ = "=0.11.2"
psycopg2-binary = "=2.9.9"
django-ninja = "^1.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
